Update rspec, eventmachine, rcov->simplecov
Dropped rake task to sh because the tasks are more brittle than the
binary interface, and ultimately it just shells out in the end, anyway.
Was able to remove a lot of the tedious parts (setting the load paths)
because they're now RSpec defaults.

Because spec is in the load path, don't need to do the requires
relative to the file, can just require spec helper directly.

The mocks in the before(:all) were blowing up because RSpec resets mocks
after each example is run. So it doesn't let you have mocks that span
the duration of the tests (at least that's what I expect the reasoning
to be). Since they didn't require any particular behaviour or interface,
I just replaced them with empty objects.

Took out the check_messages spec method, there's a better version baked
into rspec as the =~ operator

There is still one test failing:
$ rspec spec/rubarb/server_failure_spec.rb --line 66

Tried looking into it, but IDK why (it doesn't seem to raise on the no
method error, though when I explicitly raise, it suffers a "Connection
Failure"

We should probably also get rid of Jeweler, it abstracts things that are
easy to do by hand (build and push gems). And since it has the gemspec
embedded in it, inside the Rakefile, we can't gell Bundler to use our
gemspec, so we have to list all deps in both places. I was going to do this,
but couldn't figure out how we track our version, so figured I'd let
someone else.